Start using python numpy SWIG interface.
