+++ Bindings/Python/swig/python_simbody.i
@@ -16,6 +16,11 @@
 using namespace SimTK;
 %}
 
+// Add support for converting between NumPy and C arrays.
+%include "numpy.i"
+%init %{
+    import_array();
+%}
 
 %include "python_preliminaries.i"
 
@@ -39,14 +44,45 @@ using namespace SimTK;
 
 // Typemaps
 // ========
-// Allow passing python objects into OpenSim functions. For example,
-// pass a list of 3 float's into a function that takes a Vec3 as an argument.
+// Python users may feel more comfortable providing/getting SimTK numeric types
+// as NumPy types. We can use SWIG NumPy typemaps to accomplish
+// this.
+// https://docs.scipy.org/doc/numpy/reference/swig.interface-file.html
+// https://scipy-cookbook.readthedocs.io/items/SWIG_NumPy_examples.html
+// For constructor.
+// The Python version of any OpenSim function taking the pair of arguments
+// (int n, double* data) will instead take a single NumPy array.
+%apply (int DIM1, double* IN_ARRAY1) {
+    (int n, double* numpydata)
+};
+// The Python version of any Moco function taking sets of arguments
+// (int nrow, int ncol, double* data) will instead take a single
+// NumPy 2D array.
+%apply (int DIM1, int DIM2, double* IN_ARRAY2) {
+    (int nrow, int ncol, double* numpydata)
+};
+// The typemaps for functions that return matrices is more complicated: we must
+// pass in a NumPy array that already has the correct size. We create a (hidden)
+// C++ function that takes, for example, (int n, double* out)
+// (e.g., _getAsMat()).
+// Then we create a Python function that invokes the hidden C++ function with
+// a correctly-sized NumPy array (e.g., getAsMat()).
+%apply (int DIM1, double* ARGOUT_ARRAY1) {
+    (int n, double* numpyout)
+};
+// Similar to above but for 2D arrays.
+%apply (int DIM1, int DIM2, double* INPLACE_FARRAY2) {
+    (int nrow, int ncol, double* numpyout)
+};
+
+// An alternative is to use typemaps to allow OpenSim functions to accept and
+// return Python/NumPy types. For example, passing a list of 3 floats into a
+// function that takes a Vec3 as an argument.
+// The typemaps below allow this, but we opted not to go this route, because
+// it remove the ability to pass arguments using the C++ types.
+// For example, with these typemaps, Model::setGravity() can no longer take a
+// Vec3.
 /*
-TODO disabling these for now as the typemaps remove the ability to pass
-arguments using the C++ types. For example, with these typemaps,
-Model::setGravity() can no longer take a Vec3. We can revisit typemaps if we
-can keep the ability to use the original argument types.
-These typemaps work, though.
 %typemap(in) SimTK::Vec3 {
     SimTK::Vec3 v;
     if (!PySequence_Check($input)) {
@@ -126,7 +162,16 @@ These typemaps work, though.
 // Pythonic operators
 // ==================
 %extend SimTK::Vec<3> {
-     double __getitem__(int i) const {
+    SimTK::Vec(int n, double* numpydata) {
+        SimTK_ERRCHK_ALWAYS(n == 3, "Vec3()", "Size of input must be 3.");
+        return new SimTK::Vec<3>(numpydata);
+    }
+    static SimTK::Vec<3> createFromMat(int n, double* numpydata) {
+        SimTK_ERRCHK_ALWAYS(n == 3, "createFromMat()",
+                    "Size of input must be 3.");
+        return SimTK::Vec3(numpydata);
+    }
+    double __getitem__(int i) const {
         SimTK_INDEXCHECK_ALWAYS(i, $self->size(), "Vec3.__getitem__()");
         return $self->operator[](i);
     }
@@ -137,8 +182,25 @@ These typemaps work, though.
     //SimTK::Vec<3> __add__(const SimTK::Vec<3>& v) const {
     //    return *($self) + v;
     //}
+    void _getAsMat(int n, double* numpyout) const {
+        SimTK_ASSERT_ALWAYS(n == 3, "Size of input must be 3.");
+        for (int i = 0; i < n; ++i) {
+            numpyout[i] = $self->operator[](i);
+        }
+    }
+%pythoncode %{
+    def getAsMat(self):
+        return self._getAsMat(self.size())
+%};
 };
+
 %extend SimTK::Vector_<double> {
+    // SimTK::Vector_(int n, double* numpydata) {
+    //     return new SimTK::Vector_<double>(n, data, true);
+    // }
+    static SimTK::Vector_<double> createFromMat(int n, double* numpydata) {
+        return SimTK::Vector_<double>(n, numpydata, true);
+    }
     double __getitem__(int i) const {
         SimTK_INDEXCHECK_ALWAYS(i, $self->size(), "Vector.__getitem__()");
         return $self->operator[](i);
@@ -147,5 +209,14 @@ These typemaps work, though.
         SimTK_INDEXCHECK_ALWAYS(i, $self->size(), "Vector.__setitem__()");
         $self->operator[](i) = value;
     }
+    void _getAsMat(int n, double* numpyout) const {
+        SimTK_ASSERT1_ALWAYS(n == $self->size(), "Size of input must be %i.",
+                $self->size());
+        std::copy_n($self->getContiguousScalarData(), n, numpyout);
+    }
+%pythoncode %{
+    def getAsMat(self):
+        return self._getAsMat(self.size())
+%};
 };
