+++ Bindings/Python/tests/test_simbody.py
@@ -5,6 +5,8 @@
 import os
 import unittest
 
+import numpy as np
+
 import opensim as osim
 
 test_dir = os.path.join(os.path.dirname(os.path.abspath(osim.__file__)),
@@ -14,6 +16,167 @@ test_dir = os.path.join(os.path.dirname(os.path.abspath(osim.__file__)),
 osim.Model.setDebugLevel(0)
 
 class TestSimbody(unittest.TestCase):
+
+    def test_vec3_typemaps(self):
+        npv = np.array([5, 3, 6])
+        v1 = osim.Vec3(npv)
+        v2 = v1.getAsMat()
+        assert (npv == v2).all()
+
+        # Incorrect size.
+        with self.assertRaises(RuntimeError):
+            osim.Vec3(np.array([]))
+        with self.assertRaises(RuntimeError):
+            osim.Vec3(np.array([5, 1]))
+        with self.assertRaises(RuntimeError):
+            osim.Vec3(np.array([5, 1, 6, 3]))
+
+        # createFromMat()
+        npv = np.array([1, 6, 8])
+        v1 = osim.Vec3.createFromMat(npv)
+        v2 = v1.getAsMat()
+        assert (npv == v2).all()
+
+        # Incorrect size.
+        with self.assertRaises(RuntimeError):
+            osim.Vec3.createFromMat(np.array([]))
+        with self.assertRaises(RuntimeError):
+            osim.Vec3.createFromMat(np.array([5, 1]))
+        with self.assertRaises(RuntimeError):
+            osim.Vec3.createFromMat(np.array([5, 1, 6, 3]))
+
+    def test_vec3_operators(self):
+        v1 = osim.Vec3(1, 2, 3)
+        # Tests __getitem__().
+        assert v1[0] == 1
+        assert v1[1] == 2
+
+        # Out of bounds.
+        with self.assertRaises(RuntimeError):
+            v1[-1]
+        with self.assertRaises(RuntimeError):
+            v1[3]
+        with self.assertRaises(RuntimeError):
+            v1[5]
+
+        # Tests __setitem__().
+        v1[0] = 5
+        assert v1[0] == 5
+
+        # Out of bounds.
+        with self.assertRaises(RuntimeError):
+            v1[-1] = 5
+        with self.assertRaises(RuntimeError):
+            v1[3] = 1.3
+
+        # Add. TODO removed for now.
+        v2 = osim.Vec3(5, 6, 7)
+        #v3 = v1 + v2
+        #assert v3[0] == 10
+        #assert v3[1] == 8
+        #assert v3[2] == 10
+
+        # Length.
+        assert len(v2) == 3
+
+    def test_vector_typemaps(self):
+        npv = np.array([5, 3, 6, 2, 9])
+        v1 = osim.Vector.createFromMat(npv)
+        v2 = v1.getAsMat()
+        assert (npv == v2).all()
+
+        npv = np.array([])
+        v1 = osim.Vector.createFromMat(npv)
+        v2 = v1.getAsMat()
+        assert (npv == v2).all()
+
+#     def test_rowvector_typemaps(self):
+#         npv = np.array([5, 3, 6, 2, 9])
+#         v1 = osim.RowVector(npv)
+#         v2 = v1.getAsMat()
+#         assert npv == v2
+#
+#         npv = np.array([])
+#         v1 = osim.RowVector(npv)
+#         v2 = v1.getAsMat()
+#         assert npv == v2
+
+#     def test_vectorview_typemaps(self):
+
+#     def test_rowvectorview_typemaps(self):
+
+#     def test_matrix_typemaps(self):
+#         npv = np.array([[5, 3], [3, 6]])
+#         v1 = osim.Matrix(npv)
+#         v2 = v1.getAsMat()
+#         assert npv == v2
+#
+#         npv = np.array([[]])
+#         v1 = osim.Vector(npv)
+#         v2 = v1.getAsMat()
+#         assert npv == v2
+#
+#         npv = np.array([])
+#         v1 = osim.Vector(npv)
+#         v2 = v1.getAsMat()
+#         assert npv == v2
+
+    def test_vector_operators(self):
+        v = osim.Vector(5, 3)
+
+        # Tests __getitem__()
+        assert v[0] == 3
+        assert v[4] == 3
+
+        # Out of bounds.
+        with self.assertRaises(RuntimeError):
+            v[-1]
+        with self.assertRaises(RuntimeError):
+            v[5]
+
+        # Tests __setitem__()
+        v[0] = 15
+        assert v[0] == 15
+
+        with self.assertRaises(RuntimeError):
+            v[-1] = 12
+        with self.assertRaises(RuntimeError):
+            v[5] = 14
+        with self.assertRaises(RuntimeError):
+            v[9] = 18
+
+        # Size.
+        assert len(v) == 5
+
+    def test_exceptions(self):
+        with self.assertRaises(RuntimeError):
+            osim.Model("NONEXISTANT_FILE_NAME")
+        with self.assertRaises(RuntimeError):
+            m = osim.Model()
+            # Asking for the visualizer just after constructing a model
+            # throws an exception.
+            m.getVisualizer()
+
+    # def test_typemaps(self):
+    #     # TODO disabled for now
+    #     m = osim.Model()
+    #     m.setGravity(osim.Vec3(1, 2, 3))
+    #     m.setGravity([1, 2, 3])
+
+    #     with self.assertRaises(ValueError):
+    #         m.setGravity(['a', 2, 3])
+    #     with self.assertRaises(ValueError):
+    #         m.setGravity([1, 2])
+    #     with self.assertRaises(ValueError):
+    #         m.setGravity([1, 2, 6, 3])
+
+    def test_printing(self):
+        v1 = osim.Vec3(1, 3, 2)
+        assert v1.__str__() == "~[1,3,2]"
+
+        v2 = osim.Vector(7, 3)
+        assert v2.__str__() == "~[3 3 3 3 3 3 3]"
+
     def test_SimbodyMatterSubsystem(self):
         model = osim.Model(os.path.join(test_dir,
             "gait10dof18musc_subject01.osim"))
