+++ Bindings/common.i
@@ -405,6 +405,8 @@ namespace OpenSim {
 
 %include <OpenSim/Common/CSVFileAdapter.h>
 %include <OpenSim/Common/XsensDataReader.h>
+
+#ifdef WITH_BTK
 %include <OpenSim/Common/C3DFileAdapter.h>
 
 %extend OpenSim::C3DFileAdapter {
@@ -427,6 +429,7 @@ namespace OpenSim {
         $self->setLocationForForceExpression(location);
     };
 };
+#endif
 
 namespace OpenSim {
     %ignore TableSource_::TableSource_(TableSource_ &&);
